several changes (may not be compete)

+++ modules/home-manager/shell/rgr.nix
@@ -5,19 +5,18 @@
     ...
 }:
 with lib; {
-    options.pos.rgr = {
+    options.pos.shell.rgr = {
         enable = mkOption {
             type = types.bool;
             default = false;
             description = "Enable vim-like file explorer.";
         };
     };
-
-    config = mkIf (config.pos.rgr.enable
-        && config.pos.enable) {
+    config = mkIf (config.pos.shell.rgr.enable
+    && config.pos.shell.enable
+    && config.pos.enable) {
         programs.ranger = {
             enable = true;
-
             plugins = [
                 {
                     name = "devicons";
@@ -27,34 +26,36 @@ with lib; {
                     };
                 }
             ];
-
             settings = {
                 preview_images = true;
                 preview_images_method = "ueberzug";
             };
-
             extraConfig = "default_linemode devicons";
         };
 
-        # In-terminal image preview
-        home.packages = with pkgs; [
-            ueberzugpp
-        ];
+        home.packages = with pkgs; [ueberzugpp];
 
         programs.fish = {
-            shellAlias.rgr = "ranger";
-            interactiveShellInit = ''
-                function rcd
-                    set tmpfile (mktemp)
-                    ranger --choosedir=$tmpfile $argv
-                    if test -s $tmpfile
-                        set dest (cat $tmpfile)
-                        if test -d "$dest"
-                            cd "$dest"
-                        end
-                    end
-                    rm -f $tmpfile
+            shellAliases.rgr = "ranger";
+            functions.rcd = ''
+                set start_dir (test -n "$argv[1]" && echo $argv[1] || echo $PWD)
+
+                if not test -d "$start_dir"
+                    echo "rcd: not a directory: $start_dir" >&2
+                    return 1
                 end
+
+                set tmpfile (mktemp)
+                ranger --choosedir=$tmpfile $start_dir
+
+                if test -s $tmpfile
+                    set dest (cat $tmpfile)
+                    if test -d "$dest"
+                        cd "$dest"
+                    end
+                end
+
+                rm -f $tmpfile
             '';
         };
     };
